Revert "refactor(ui): estructura de carpetas"
