+++ src/app/app.routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from './features/home/home.component';
+import { HomeComponent } from '../features/home/home.component';
 import { LayoutComponent } from './layout/layout.component';
 
 export const routes: Routes = [
