+++ src/shared/components/footer/footer.component.ts
