+++ src/shared/components/loader/loader.component.ts
